Move more modules from flavours/glitch/utils to flavours/glitch
