+++ app/javascript/flavours/glitch/compare_id.js
