+++ app/javascript/flavours/glitch/extra_polyfills.js
