+++ app/javascript/flavours/glitch/is_mobile.js
