+++ app/javascript/flavours/glitch/load_keyboard_extensions.js
