+++ app/javascript/flavours/glitch/load_polyfills.js
