+++ app/javascript/flavours/glitch/main.js
@@ -4,7 +4,7 @@ import { setupBrowserNotifications } from 'flavours/glitch/actions/notifications
 import Mastodon, { store } from 'flavours/glitch/containers/mastodon';
 import ready from 'flavours/glitch/ready';
 
-const perf = require('flavours/glitch/utils/performance');
+const perf = require('flavours/glitch/performance');
 
 /**
  * @returns {Promise<void>}
