+++ app/javascript/flavours/glitch/packs/home.js
@@ -1,8 +1,8 @@
 import 'packs/public-path';
-import loadPolyfills from 'flavours/glitch/utils/load_polyfills';
+import loadPolyfills from 'flavours/glitch/load_polyfills';
 
 loadPolyfills().then(async () => {
-  const { default: main } = await import('flavours/glitch/utils/main');
+  const { default: main } = await import('flavours/glitch/main');
 
   return main();
 }).catch(e => {
