+++ app/javascript/flavours/glitch/packs/settings.js
@@ -1,7 +1,7 @@
 import 'packs/public-path';
-import loadPolyfills from 'flavours/glitch/utils/load_polyfills';
+import loadPolyfills from 'flavours/glitch/load_polyfills';
 import ready from 'flavours/glitch/ready';
-import loadKeyboardExtensions from 'flavours/glitch/utils/load_keyboard_extensions';
+import loadKeyboardExtensions from 'flavours/glitch/load_keyboard_extensions';
 import 'cocoon-js-vanilla';
 
 function main() {
