+++ app/javascript/flavours/glitch/performance.js
