+++ app/javascript/flavours/glitch/stream.js
