+++ app/javascript/flavours/glitch/uuid.js
